+++ docs/CHANGELOG.md
@@ -14,6 +14,10 @@ nav_order: 6
 
     *Joel Hawksley*
 
+* Add `SystemSpecHelpers` for use with RSpec.
+
+    *Meyric Rawlings*, *Joel Hawksley*
+
 ## 4.0.0.alpha7
 
 * BREAKING: Remove deprecated `use_helper(s)`. Use `include MyHelper` or `helpers.` proxy instead.